refactor: view 패키지를 console 하위로 이동
